requirements.txt
flask
flask_login
flask_wtf
wtforms
sqlalchemy_serializer
flask_restful
python-dotenv